pyproject.toml
[project]
name = "linker"
version = "1.0"
description = ""
authors = [
    {name = "None"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "fastapi (>=0.127.0,<0.128.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "beanie (>=2.0.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "qdrant-client (>=1.16.2,<2.0.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
